Move defaults to common file
