+++ api_tests/src/actors/actor.ts
@@ -16,15 +16,18 @@ import "../setup_chai";
 import { Asset, AssetKind, toKey, toKind } from "../asset";
 import { CndInstance } from "../cnd/cnd_instance";
 import { Ledger, LedgerKind } from "../ledgers/ledger";
-import { HarnessGlobal, LedgerConfig, sleep } from "../utils";
+import { LedgerConfig, sleep } from "../utils";
 import { Wallet, Wallets } from "../wallets";
 import { Actors } from "./index";
 import { LndInstance } from "../ledgers/lnd_instance";
 import { sha256 } from "js-sha256";
 import { InvoiceState } from "@radar/lnrpc";
-import { parseEther } from "ethers/utils";
-
-declare var global: HarnessGlobal;
+import {
+    defaultAssetDescription,
+    defaultExpiryTimes,
+    defaultLedgerDescriptionForLedger,
+    defaultLedgerKindForAsset,
+} from "./defaults";
 
 export class Actor {
     public static defaultActionConfig = {
@@ -965,79 +968,4 @@ export class Actor {
             );
         }
     }
-}
-
-function defaultLedgerKindForAsset(asset: AssetKind): LedgerKind {
-    switch (asset) {
-        case AssetKind.Bitcoin:
-            return LedgerKind.Bitcoin;
-        case AssetKind.Ether:
-            return LedgerKind.Ethereum;
-        case AssetKind.Erc20:
-            return LedgerKind.Ethereum;
-    }
-}
-
-/**
- * WIP as the cnd REST API routes for lightning are not yet defined.
- * @param ledger
- * @returns The ledger formatted as needed for the request body to cnd HTTP API on the lightning route.
- */
-function defaultLedgerDescriptionForLedger(ledger: LedgerKind): Ledger {
-    switch (ledger) {
-        case LedgerKind.Lightning: {
-            return {
-                name: LedgerKind.Lightning,
-            };
-        }
-        case LedgerKind.Bitcoin: {
-            return {
-                name: LedgerKind.Bitcoin,
-                network: "regtest",
-            };
-        }
-        case LedgerKind.Ethereum: {
-            return {
-                name: LedgerKind.Ethereum,
-                chain_id: 17,
-            };
-        }
-    }
-}
-
-function defaultAssetDescription(asset: AssetKind, ledger: LedgerKind): Asset {
-    switch (asset) {
-        case AssetKind.Bitcoin: {
-            return {
-                name: AssetKind.Bitcoin,
-                ledger,
-                quantity: "10000000",
-            };
-        }
-        case AssetKind.Ether: {
-            return {
-                name: AssetKind.Ether,
-                ledger,
-                quantity: parseEther("10").toString(),
-            };
-        }
-        case AssetKind.Erc20: {
-            return {
-                name: AssetKind.Erc20,
-                ledger,
-                quantity: parseEther("100").toString(),
-                token_contract: global.tokenContract,
-            };
-        }
-    }
-}
-
-function defaultExpiryTimes() {
-    const alphaExpiry = Math.round(Date.now() / 1000) + 8;
-    const betaExpiry = Math.round(Date.now() / 1000) + 3;
-
-    return {
-        alpha_expiry: alphaExpiry,
-        beta_expiry: betaExpiry,
-    };
 }
